fix: separate client component and metadata config in user homepage
